moved common methods to single connector
