+++ device-connectors/src/testflinger_device_connectors/devices/maas2/__init__.py
@@ -39,7 +39,7 @@ class DeviceConnector(DefaultDevice):
 
         super().provision(args)
 
-        self._disconnect_usb_stick(config)
+        ZapperConnector.disconnect_usb_stick(config)
 
         device = Maas2(args.config, args.job_data)
         logger.info("BEGIN provision")
@@ -58,24 +58,6 @@ class DeviceConnector(DefaultDevice):
         finally:
             serial_proc.stop()
             logger.info("END provision")
-
-    def _disconnect_usb_stick(self, config: dict) -> None:
-        """Try to disconnect the USB stick.
-
-        This is a non-blocking operation - if the Zapper is not available,
-        we simply skip this step.
-        """
-        control_host = config.get("control_host")
-        if not control_host:
-            return
-
-        try:
-            ZapperConnector.wait_ready(control_host)
-            ZapperConnector.typecmux_set_state(control_host, "OFF")
-        except (TimeoutError, ConnectionError, Exception) as e:
-            logger.debug(
-                "Could not disconnect USB stick on %s: %s", control_host, e
-            )
 
     def cleanup(self, args):
         device = Maas2(args.config, args.job_data)
