+++ device-connectors/src/testflinger_device_connectors/devices/oem_autoinstall/tests/test_zapper_oem.py
@@ -185,75 +185,3 @@ class ZapperOemTests(unittest.TestCase):
 
 if __name__ == "__main__":
     unittest.main()
-
-
-class TestOemAutoinstallDisconnectUsbStick:
-    """Tests for oem_autoinstall USB stick disconnect functionality."""
-
-    def test_disconnect_usb_stick_success(self, mocker):
-        """Test _disconnect_usb_stick succeeds when Zapper is available."""
-        from testflinger_device_connectors.devices.oem_autoinstall import (
-            DeviceConnector,
-        )
-        from testflinger_device_connectors.devices.zapper import (
-            ZapperConnector,
-        )
-
-        config = {"control_host": "zapper-host", "device_ip": "1.2.3.4"}
-
-        mock_wait_ready = mocker.patch.object(ZapperConnector, "wait_ready")
-        mock_typecmux = mocker.patch.object(
-            ZapperConnector, "typecmux_set_state"
-        )
-
-        connector = DeviceConnector(config)
-        connector._disconnect_usb_stick(config)
-
-        mock_wait_ready.assert_called_once_with("zapper-host")
-        mock_typecmux.assert_called_once_with("zapper-host", "OFF")
-
-    def test_disconnect_usb_stick_no_control_host(self, mocker):
-        """Test _disconnect_usb_stick skips when no control_host."""
-        from testflinger_device_connectors.devices.oem_autoinstall import (
-            DeviceConnector,
-        )
-        from testflinger_device_connectors.devices.zapper import (
-            ZapperConnector,
-        )
-
-        config = {"device_ip": "1.2.3.4"}
-
-        mock_wait_ready = mocker.patch.object(ZapperConnector, "wait_ready")
-        mock_typecmux = mocker.patch.object(
-            ZapperConnector, "typecmux_set_state"
-        )
-
-        connector = DeviceConnector(config)
-        connector._disconnect_usb_stick(config)
-
-        mock_wait_ready.assert_not_called()
-        mock_typecmux.assert_not_called()
-
-    def test_disconnect_usb_stick_timeout_non_blocking(self, mocker):
-        """Test _disconnect_usb_stick handles timeout gracefully."""
-        from testflinger_device_connectors.devices.oem_autoinstall import (
-            DeviceConnector,
-        )
-        from testflinger_device_connectors.devices.zapper import (
-            ZapperConnector,
-        )
-
-        config = {"control_host": "zapper-host", "device_ip": "1.2.3.4"}
-
-        mocker.patch.object(
-            ZapperConnector, "wait_ready", side_effect=TimeoutError
-        )
-        mock_typecmux = mocker.patch.object(
-            ZapperConnector, "typecmux_set_state"
-        )
-
-        connector = DeviceConnector(config)
-        # Should not raise - non-blocking operation
-        connector._disconnect_usb_stick(config)
-
-        mock_typecmux.assert_not_called()
